+++ clinica-odontologica/src/main/java/com/dh/clinica/service/PacienteService.java
@@ -1,6 +1,7 @@
 package com.dh.clinica.service;
 
 
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Paciente;
 import com.dh.clinica.repository.PacienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,7 +9,9 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
+// TODO: este servicio debería implementar una interface que defina el contrato de los métodos que se usarán
 @Service
 public class PacienteService {
 
@@ -32,7 +35,11 @@ public class PacienteService {
         return pacienteRepository.findAll();
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(Long id) throws ResourceNotFoundException {
+        // Librería Objects permite hacer validaciones tratando nulos
+        if(Objects.isNull(buscar(id))) {
+            throw new ResourceNotFoundException("No existe paciente con id: " + id);
+        }
         pacienteRepository.deleteById(id);
     }
