+++ clinica-odontologica/src/main/java/com/dh/clinica/service/TurnoService.java
@@ -1,5 +1,7 @@
 package com.dh.clinica.service;
 
+import com.dh.clinica.exceptions.BadRequestException;
+import com.dh.clinica.exceptions.ResourceNotFoundException;
 import com.dh.clinica.model.Odontologo;
 import com.dh.clinica.model.Paciente;
 import com.dh.clinica.model.Turno;
@@ -11,7 +13,10 @@ import org.springframework.stereotype.Service;
 
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
 
+// TODO: este servicio debería implementar una interface que defina el contrato de los métodos que se usarán
 @Service
 public class TurnoService {
 
@@ -28,11 +33,21 @@ public class TurnoService {
         this.pacienteRepository = pacienteRepository;
     }
 
-    public Turno registrarTurno(Turno turno) {
-        Odontologo odontologo = odontologoRepository.findById(turno.getOdontologo().getId()).get();
-        turno.setOdontologo(odontologo);
-        Paciente paciente = pacienteRepository.findById(turno.getPaciente().getId()).get();
-        turno.setPaciente(paciente);
+    public Turno registrarTurno(Turno turno) throws BadRequestException {
+
+        Optional<Odontologo> odontologo = odontologoRepository.findById(turno.getOdontologo().getId());
+
+        Optional<Paciente> paciente = pacienteRepository.findById(turno.getPaciente().getId());
+
+
+        if (paciente.isEmpty() || odontologo.isEmpty()) {
+            throw new BadRequestException("El paciente o el odontologo no existen");
+        }
+
+        turno.setOdontologo(odontologo.get());
+        turno.setPaciente(paciente.get());
+
+
         turno.setDate(new Date());
         return turnoRepository.save(turno);
     }
@@ -41,7 +56,11 @@ public class TurnoService {
         return turnoRepository.findAll();
     }
 
-    public void eliminar(Long id) {
+    public void eliminar(Long id) throws ResourceNotFoundException {
+        // Librería Objects permite hacer validaciones tratando nulos
+        if(Objects.isNull(buscar(id))) {
+            throw new ResourceNotFoundException("No existe turno con id: " + id);
+        }
         turnoRepository.deleteById(id);
     }
